backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
bcrypt==4.0.1
reportlab
httpx
psycopg2-binary
python-dotenv
supabase==2.6.0
mangum
